requirements.txt
dearpygui==1.8.0
Pillow==9.3.0
pyperclip==1.8.2
pytesseract==0.3.10
